Finished frame scaling.
Changed the BBox to use doubles instead of ints to work with scaling
frames.

+++ src/BBox.java
@@ -1,5 +1,5 @@
 /**
- * Class for a bounding box just holds 4 ints
+ * Class for a bounding box just holds 4 doubles
  *
  * @x1: the xth pixel of the frame for the first corner
  * @y1: the yth pixel of the frame for the first corner
@@ -11,40 +11,40 @@
  */
 public class BBox {
 
-    private int x1 = 0;
-    private int y1 = 0;
-    private int x2 = 0;
-    private int y2 = 0;
+    private double x1 = 0;
+    private double y1 = 0;
+    private double x2 = 0;
+    private double y2 = 0;
 
-    public int x1() {
+    public double x1() {
         return this.x1;
     }
 
-    public int y1() {
+    public double y1() {
         return this.y1;
     }
 
-    public int x2() {
+    public double x2() {
         return this.x2;
     }
 
-    public int y2() {
+    public double y2() {
         return this.y2;
     }
 
-    public void setx1(int x) {
+    public void setx1(double x) {
         this.x1 = x;
     }
 
-    public void sety1(int y) {
+    public void sety1(double y) {
         this.y1 = y;
     }
 
-    public void setx2(int x) {
+    public void setx2(double x) {
         this.x2 = x;
     }
 
-    public void sety2(int y) {
+    public void sety2(double y) {
         this.y2 = y;
     }
 
@@ -55,7 +55,7 @@ public class BBox {
 
     }
 
-    public BBox(int x1, int y1, int x2, int y2) {
+    public BBox(double x1, double y1, double x2, double y2) {
         this.x1 = x1;
         this.y1 = y1;
         this.x2 = x2;
